+++ CommonWallet/Classes/Modules/WithdrawAmount/ViewModel/WithdrawAmountViewModelFactory.swift
@@ -65,7 +65,7 @@ extension WithdrawAmountViewModelFactory: WithdrawAmountViewModelFactoryProtocol
         return AmountInputViewModel(amount: amount,
                                     limit: amountLimit,
                                     formatter: localizedFormatter,
-                                    precision: UInt8(localizedFormatter.maximumFractionDigits))
+                                    precision: Int16(localizedFormatter.maximumFractionDigits))
     }
 
     func createDescriptionViewModel() throws -> DescriptionInputViewModel {
